refactor: encapsulate input handler. make processor independant.
