+++ core/src/main/java/dev/kingdomino/game/DraftInputProcessor.java
@@ -1,145 +1,36 @@
 package dev.kingdomino.game;
 
-import com.badlogic.gdx.InputProcessor;
 import com.badlogic.gdx.Input.Keys;
 
-import dev.kingdomino.game.Event.TriggerType;
+import dev.kingdomino.game.DraftInputHandler.Action;
 
-public class DraftInputProcessor implements InputProcessor {
-    private EventManager eventManager = EventManager.getInstance();
-    private GameManager gameManager;
-    private Turn nextTurn;
-    public int selectionIndex;
-    public int remainingDrafts;
+public class DraftInputProcessor extends AbstractInputProcessor {
+    private final DraftInputHandler draftInputHandler;
 
-    public boolean updated;
-    public boolean exit;
-    public boolean show;
-
-    public DraftInputProcessor(GameManager gm) {
-        this.gameManager = gm;
-        this.updated = true;
-        this.exit = false;
-        this.show = true;
-    }
-
-    public void reset() {
-        nextTurn = gameManager.getNextTurn();
-        remainingDrafts = nextTurn.getDraft().length;
-        updated = true;
-        exit = false;
-        show = true;
-        
-        selectionIndex = 0;
-        while (nextTurn.isSelected(selectionIndex)) {
-            selectionIndex = (selectionIndex + 1) % remainingDrafts;
-        }
-    }
-
-    public int getSelectionIndex() {
-        return this.selectionIndex;
+    public DraftInputProcessor(DraftInputHandler draftInputHandler) {
+        this.draftInputHandler = draftInputHandler;
     }
 
     @Override
     public boolean keyDown(int keycode) {
-        if (gameManager.getCurrentState() == GameManager.GameState.TURN_START) {
-            show = false;
-            return true;
-        }
-        if (gameManager.getCurrentState() != GameManager.GameState.TURN_CHOOSING) {
-            updated = true;
-            return false;
-        }
-        Event e = null;
+        return draftInputHandler.keyDown(translateKeycodeToAction(keycode));
+    }
 
+    private Action translateKeycodeToAction(int keycode) {
         switch (keycode) {
             case Keys.LEFT:
             case Keys.UP:
             case Keys.W:
-                while (true) {
-                    selectionIndex = (selectionIndex + remainingDrafts - 1) % remainingDrafts;
-
-                    if (!nextTurn.isSelected(selectionIndex)) break;
-                }
-                
-                updated = true;
-                break;
+                return Action.MOVE_UP;
             case Keys.RIGHT:
             case Keys.DOWN:
             case Keys.S:
-                while (true) {
-                    selectionIndex = (selectionIndex + 1) % remainingDrafts;
-
-                    if (!nextTurn.isSelected(selectionIndex)) break;
-                }
-
-                updated = true;
-                break;
+                return Action.MOVE_DOWN;
             case Keys.ENTER:
             case Keys.X:
-                e = new Event(
-                        TriggerType.IMMEDIATE,
-                        false,
-                        true,
-                        null,
-                        () -> {
-                            gameManager.selectDomino(selectionIndex);
-                            exit = true;
-                        },
-                        null,
-                        null,
-                        null);
-                break;
+                return Action.SELECT_DOMINO;
             default:
-                break;
+                return Action.NONE;
         }
-
-        if (e != null) {
-            eventManager.addEvent(e, "base", false);
-            updated = true;
-        }
-
-        return true;
-
-    }
-
-    @Override
-    public boolean keyUp(int keycode) {
-        return false;
-    }
-
-    @Override
-    public boolean keyTyped(char character) {
-        return false;
-    }
-
-    @Override
-    public boolean touchDown(int screenX, int screenY, int pointer, int button) {
-        return false;
-    }
-
-    @Override
-    public boolean touchUp(int screenX, int screenY, int pointer, int button) {
-        return false;
-    }
-
-    @Override
-    public boolean touchCancelled(int screenX, int screenY, int pointer, int button) {
-        return false;
-    }
-
-    @Override
-    public boolean touchDragged(int screenX, int screenY, int pointer) {
-        return false;
-    }
-
-    @Override
-    public boolean mouseMoved(int screenX, int screenY) {
-        return false;
-    }
-
-    @Override
-    public boolean scrolled(float amountX, float amountY) {
-        return false;
     }
 }
